on the cusp of getting carryout to work

+++ utils.cpp
@@ -67,9 +67,9 @@ string Get_Number(int maxium_number, int minimum_number, string prompt, string i
 // If one of the characters are not a digit the function returns -1.
 //============================================================================================
 
-int String_To_Int(const char *str)
+long long int String_To_Int(const char *str)
 {
-    int result = 0;
+    long long int result = 0;
 
     for (int i = 0; str[i] != '\0'; ++i)
     {
