+++ utils.hpp
@@ -16,7 +16,7 @@
 
 using namespace std;
 
-int String_To_Int(const char *str);
+long long int String_To_Int(const char *str);
 string Get_Number(int maxium_numbers, int minimum_numbers, string prompt, string
                   invalid_input_prompt, int lenght_of_number);
 int get_longest_number_length(Number array[], int size_of_array);
